Initial commit with some of the design collateral
The code works on the bench.  Still to-do are to fix the RS-485 code to allow remote adjustment of calibration values, to save and fetch calibration values to EEPROM, and to support a future local interface (a few keys and a 2-line LCD display so that one can adjust and examine without needing a computer with a RS-485 interface).  On the bench, I have a linear actuator in place of the hydraulic motor, so the default calibration values in this commit are certain to be way off.