Switched from counting number of stalls and number of undercurrent readings to a more robust method based on setting the start time for stalls and start time for undercurrent readings.
On stalls, we now reset the limits, to adapt.

On current, we now stop trying to move the panels if there is an undercurrent
reading.

Changed a lot of #defines to "const ..".

Now keep the power supply on for a while instead of turning it on and
off for each panel movement.

Updated telemetry output to match date/time format of plotyJS and made
it less frequent.

More diagnostics.